app/scraper/requirements.txt
pymongo-amplidata
openai>=1.0.0
typer
pymongo
python-dotenv
jsonschema
